Big commit: updated gitignore file to ignore all pycache files, .vscode, and venv, updated requirements.txt to include black, refactored views based on categories
