+++ api/consumers.py
@@ -5,7 +5,12 @@ import random
 from channels.generic.websocket import AsyncWebsocketConsumer
 from django.utils import timezone
 
-from api.models import QuizSession, QuizSessionStudent, UserResponse, QuestionMultipleChoice
+from api.models import (
+    QuizSession,
+    QuizSessionStudent,
+    UserResponse,
+    QuestionMultipleChoice,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -25,7 +30,9 @@ class QuizSessionInstructorConsumer(AsyncWebsocketConsumer):
         uc = await self.fetch_user_count()
 
         await self.send(
-            text_data=json.dumps({"type": "settings", "settings": settings, "user_count": uc})
+            text_data=json.dumps(
+                {"type": "settings", "settings": settings, "user_count": uc}
+            )
         )
 
     @database_sync_to_async
@@ -102,7 +109,9 @@ class QuizSessionInstructorConsumer(AsyncWebsocketConsumer):
     @database_sync_to_async
     def delete_student_from_db(self, username):
         session = QuizSession.objects.get(code=self.code)
-        student = QuizSessionStudent.objects.get(quiz_session=session, username=username)
+        student = QuizSessionStudent.objects.get(
+            quiz_session=session, username=username
+        )
         student.delete()
         return {"status": "success", "username": username}
 
@@ -118,7 +127,9 @@ class QuizSessionInstructorConsumer(AsyncWebsocketConsumer):
         question_data = await self.fetch_current_question()
         if question_data:
             await self.send(
-                text_data=json.dumps({"type": "current_question", "question": question_data})
+                text_data=json.dumps(
+                    {"type": "current_question", "question": question_data}
+                )
             )
 
     @database_sync_to_async
@@ -148,7 +159,9 @@ class QuizSessionInstructorConsumer(AsyncWebsocketConsumer):
         question_data = await self.fetch_next_question()
         if question_data:
             await self.send(
-                text_data=json.dumps({"type": "next_question", "question": question_data})
+                text_data=json.dumps(
+                    {"type": "next_question", "question": question_data}
+                )
             )
         else:
             print("Ending Quiz")
@@ -173,10 +186,14 @@ class QuizSessionInstructorConsumer(AsyncWebsocketConsumer):
             print("Failed to end the quiz; session not found.")
 
     async def start_quiz(self):
-        await self.channel_layer.group_send(f"quiz_session_{self.code}", {"type": "quiz_started"})
+        await self.channel_layer.group_send(
+            f"quiz_session_{self.code}", {"type": "quiz_started"}
+        )
 
         await self.send(
-            text_data=json.dumps({"type": "quiz_started", "message": "Quiz has started!"})
+            text_data=json.dumps(
+                {"type": "quiz_started", "message": "Quiz has started!"}
+            )
         )
 
     async def student_joined(self, event):
@@ -198,7 +215,9 @@ class QuizSessionInstructorConsumer(AsyncWebsocketConsumer):
 
     async def user_response(self, event):
         await self.send(
-            text_data=json.dumps({"type": "user_response", "response": event["response"]})
+            text_data=json.dumps(
+                {"type": "user_response", "response": event["response"]}
+            )
         )
 
 
@@ -266,7 +285,9 @@ class StudentConsumer(AsyncWebsocketConsumer):
         try:
             session = QuizSession.objects.get(code=code)
             # studentUser = Student.objects.get(user_id=user_id)
-            student = QuizSessionStudent.objects.create(username=username, quiz_session=session)
+            student = QuizSessionStudent.objects.create(
+                username=username, quiz_session=session
+            )
             return {
                 "status": "success",
                 "message": "Student created successfully",
@@ -323,7 +344,9 @@ class StudentConsumer(AsyncWebsocketConsumer):
             if session_settings.get("skip_question_logic") == "streak":
                 if (
                     student.skip_count < session_settings.get("skip_count_per_student")
-                    and correct_responses % session_settings.get("skip_question_streak_count") == 0
+                    and correct_responses
+                    % session_settings.get("skip_question_streak_count")
+                    == 0
                 ):
                     grant_response = await self.grant_skip_power_up(student_id)
                     if grant_response.get("status") == "success":
@@ -385,7 +408,10 @@ class StudentConsumer(AsyncWebsocketConsumer):
                 await self.increment_skip_count(student.get("id"))
                 await self.send(
                     text_data=json.dumps(
-                        {"type": "skip_power_up_used", "message": "Question skipped successfully."}
+                        {
+                            "type": "skip_power_up_used",
+                            "message": "Question skipped successfully.",
+                        }
                     )
                 )
         else:
