+++ api/tests/tests.py
@@ -105,18 +105,22 @@ class InstructorViewTest(BaseTest):
 
     def test_get_instructor(self):
         url = reverse(
-            "instructor-detail", kwargs={"instructor_id": self.new_user_instructor.instructor.id}
+            "instructor-detail",
+            kwargs={"instructor_id": self.new_user_instructor.instructor.id},
         )
         response = self.client_instructor.get(url)
 
         self.assertEqual(response.status_code, 200)
 
         response_data = response.json()
-        self.assertEqual(response_data["instructor"]["id"], self.new_user_instructor.instructor.id)
+        self.assertEqual(
+            response_data["instructor"]["id"], self.new_user_instructor.instructor.id
+        )
 
     def test_put_instructor(self):
         url = reverse(
-            "instructor-detail", kwargs={"instructor_id": self.new_user_instructor.instructor.id}
+            "instructor-detail",
+            kwargs={"instructor_id": self.new_user_instructor.instructor.id},
         )
         data = {
             "instructor": {},
@@ -130,12 +134,15 @@ class InstructorViewTest(BaseTest):
         response_data = response.json()
         self.assertEqual(response_data["message"], "Instructor updated successfully")
 
-        new_instructor = Instructor.objects.get(id=self.new_user_instructor.instructor.id)
+        new_instructor = Instructor.objects.get(
+            id=self.new_user_instructor.instructor.id
+        )
         self.assertEqual(new_instructor.user.username, data["user"]["username"])
 
     def test_delete_instructor(self):
         url = reverse(
-            "instructor-detail", kwargs={"instructor_id": self.new_user_instructor.instructor.id}
+            "instructor-detail",
+            kwargs={"instructor_id": self.new_user_instructor.instructor.id},
         )
         response = self.client_instructor.delete(url)
 
@@ -163,16 +170,22 @@ class StudentViewTest(BaseTest):
         self.assertEqual(new_student.user.username, data["user"]["username"])
 
     def test_get_student(self):
-        url = reverse("student-detail", kwargs={"student_id": self.new_user_student.student.id})
+        url = reverse(
+            "student-detail", kwargs={"student_id": self.new_user_student.student.id}
+        )
         response = self.client_student.get(url)
 
         self.assertEqual(response.status_code, 200)
 
         response_data = response.json()
-        self.assertEqual(response_data["student"]["id"], self.new_user_student.student.id)
+        self.assertEqual(
+            response_data["student"]["id"], self.new_user_student.student.id
+        )
 
     def test_put_student(self):
-        url = reverse("student-detail", kwargs={"student_id": self.new_user_student.student.id})
+        url = reverse(
+            "student-detail", kwargs={"student_id": self.new_user_student.student.id}
+        )
         data = {
             "student": {},
             "user": {
@@ -190,7 +203,9 @@ class StudentViewTest(BaseTest):
         self.assertEqual(updated_student.user.username, data["user"]["username"])
 
     def test_delete_student(self):
-        url = reverse("student-detail", kwargs={"student_id": self.new_user_student.student.id})
+        url = reverse(
+            "student-detail", kwargs={"student_id": self.new_user_student.student.id}
+        )
         response = self.client_student.delete(url)
 
         self.assertEqual(response.status_code, 200)
@@ -227,7 +242,9 @@ class QuizViewTest(BaseTest):
         response_data = response.json()
         self.assertEqual(response_data["quiz"]["id"], self.new_quiz.id)
         self.assertEqual(response_data["quiz"]["title"], self.new_quiz.title)
-        self.assertEqual(response_data["quiz"]["instructor_id"], self.new_quiz.instructor.id)
+        self.assertEqual(
+            response_data["quiz"]["instructor_id"], self.new_quiz.instructor.id
+        )
 
     def test_put_quiz(self):
         url = reverse("quiz-detail", kwargs={"quiz_id": self.new_quiz.id})
@@ -245,7 +262,9 @@ class QuizViewTest(BaseTest):
 
         updated_quiz = Quiz.objects.get(id=self.new_quiz.id)
         self.assertEqual(updated_quiz.title, data["title"])
-        self.assertEqual(updated_quiz.start_time.isoformat(), "2023-01-03T00:00:00+00:00")
+        self.assertEqual(
+            updated_quiz.start_time.isoformat(), "2023-01-03T00:00:00+00:00"
+        )
         self.assertEqual(updated_quiz.end_time.isoformat(), "2023-01-04T00:00:00+00:00")
 
     def test_delete_quiz(self):
@@ -276,9 +295,13 @@ class QuestionViewTest(BaseTest):
         response_data = response.json()
         self.assertEqual(response_data["message"], "Question created successfully")
 
-        new_question = QuestionMultipleChoice.objects.get(id=response_data["question_id"])
+        new_question = QuestionMultipleChoice.objects.get(
+            id=response_data["question_id"]
+        )
         self.assertEqual(new_question.question_text, data["question_text"])
-        self.assertEqual(new_question.incorrect_answer_list, data["incorrect_answer_list"])
+        self.assertEqual(
+            new_question.incorrect_answer_list, data["incorrect_answer_list"]
+        )
         self.assertEqual(new_question.correct_answer, data["correct_answer"])
         self.assertEqual(new_question.points, data["points"])
         self.assertEqual(new_question.quiz.id, data["quiz_id"])
@@ -299,13 +322,18 @@ class QuestionViewTest(BaseTest):
             self.new_question.incorrect_answer_list,
         )
         self.assertEqual(
-            response_data["questions"]["correct_answer"], self.new_question.correct_answer
+            response_data["questions"]["correct_answer"],
+            self.new_question.correct_answer,
         )
         self.assertEqual(response_data["questions"]["points"], self.new_question.points)
-        self.assertEqual(response_data["questions"]["quiz_id"], self.new_question.quiz.id)
+        self.assertEqual(
+            response_data["questions"]["quiz_id"], self.new_question.quiz.id
+        )
 
     def test_get_all_questions(self):
-        all_questions_url = reverse("all-questions", kwargs={"quiz_id": self.new_quiz.id})
+        all_questions_url = reverse(
+            "all-questions", kwargs={"quiz_id": self.new_quiz.id}
+        )
         response = self.client_instructor.get(all_questions_url)
 
         self.assertEqual(response.status_code, 200)
@@ -364,7 +392,9 @@ class QuestionViewTest(BaseTest):
 
         updated_question = QuestionMultipleChoice.objects.get(id=self.new_question.id)
         self.assertEqual(updated_question.question_text, data["question_text"])
-        self.assertEqual(updated_question.incorrect_answer_list, data["incorrect_answer_list"])
+        self.assertEqual(
+            updated_question.incorrect_answer_list, data["incorrect_answer_list"]
+        )
         self.assertEqual(updated_question.correct_answer, data["correct_answer"])
         self.assertEqual(updated_question.points, data["points"])
         self.assertEqual(updated_question.quiz.id, data["quiz_id"])
@@ -397,7 +427,9 @@ class UserResponseViewTest(BaseTest):
         self.assertEqual(response_data["message"], "User response created successfully")
 
     def test_put_user_response(self):
-        url = reverse("user-response-detail", kwargs={"response_id": self.new_user_response.id})
+        url = reverse(
+            "user-response-detail", kwargs={"response_id": self.new_user_response.id}
+        )
         data = {"student_id": self.new_quiz_session_student.id, "selected_answer": "1"}
         response = self.client.put(url, data, format="json")
 
@@ -422,7 +454,9 @@ class QuizSessionResultsTest(BaseTest):
         results = response.json()["results"]
 
         self.assertEqual(len(results), 1)
-        self.assertEqual(results[0]["student_username"], self.new_quiz_session_student.username)
+        self.assertEqual(
+            results[0]["student_username"], self.new_quiz_session_student.username
+        )
         self.assertEqual(results[0]["correct_answers"], 1)
         self.assertEqual(results[0]["total_questions"], 1)
