+++ hice_backend/urls.py
@@ -1,6 +1,9 @@
 from django.contrib import admin
 from django.urls import path
-from api.views import *
+from api.views.question_views import *
+from api.views.quiz_views import *
+from api.views.session_views import *
+from api.views.user_views import *
 from api.consumers import *
 
 from drf_spectacular.views import SpectacularAPIView, SpectacularSwaggerView
