+++ manage.py
@@ -9,7 +9,7 @@ dotenv.load_dotenv()
 
 def main():
     """Run administrative tasks."""
-    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'hice_backend.settings')
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "hice_backend.settings")
     try:
         from django.core.management import execute_from_command_line
     except ImportError as exc:
@@ -21,5 +21,5 @@ def main():
     execute_from_command_line(sys.argv)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     main()
