+++ requirements.txt
@@ -54,3 +54,4 @@ typing_extensions==4.12.2
 uritemplate==4.1.1
 urllib3==1.26.20
 zope.interface==6.4.post2
+black==24.4.2
